+++ src/main/java/me/lucaspickering/terraingen/world/WorldHelper.java
@@ -21,8 +21,8 @@ public class WorldHelper {
      */
     @NotNull
     public static Point tileToPixel(@NotNull World world, @NotNull TilePoint tile) {
-        final float x = world.getTileWidth() * tile.x() * 0.75f;
-        final float y = -world.getTileHeight() * (tile.x() / 2.0f + tile.y());
+        final double x = world.getTileWidth() * tile.x() * 0.75f;
+        final double y = -world.getTileHeight() * (tile.x() / 2.0f + tile.y());
         return world.getWorldCenter().plus((int) x, (int) y);
     }
 
@@ -41,10 +41,10 @@ public class WorldHelper {
     public static TilePoint pixelToTile(@NotNull World world, @NotNull Point pos) {
         final Point shiftedPos = pos.minus(world.getWorldCenter());
         // Convert it to a fractional tile point
-        final float fracX = shiftedPos.x() * 4f / 3f / world.getTileWidth();
-        final float fracY = -(shiftedPos.x() + (float) Math.sqrt(3) * shiftedPos.y())
+        final double fracX = shiftedPos.x() * 4f / 3f / world.getTileWidth();
+        final double fracY = -(shiftedPos.x() + Math.sqrt(3) * shiftedPos.y())
                             / (world.getTileRadius() * 3f);
-        final float fracZ = -fracX - fracY; // We'll need this later
+        final double fracZ = -fracX - fracY; // We'll need this later
 
         // Return the rounded point
         return TilePoint.roundPoint(fracX, fracY, fracZ);
